requirements.txt
streamlit~=1.33.0
PyPDF2~=3.0.1
langchain~=0.1.16
python-dotenv~=1.0.1